Split off library
